+++ internal/imap/mailbox_messages.go
@@ -383,6 +383,12 @@ func (im *imapMailbox) ListMessages(isUID bool, seqSet *imap.SeqSet, items []ima
 		im.panicHandler.HandlePanic()
 	}()
 
+	// EXPUNGE cannot be sent during listing and can come only from
+	// the event loop, so we prevent any server side update to avoid
+	// the problem.
+	im.storeUser.PauseEventLoop(true)
+	defer im.storeUser.PauseEventLoop(false)
+
 	var markAsReadIDs []string
 	markAsReadMutex := &sync.Mutex{}